Add files via upload
Implementation of a Federated Learning (FL) client using the Flower (FLwr) framework, with a focus on a binary classification task related to heart disease prediction.

The Server.py script serves as a client in a Federated Learning (FL) system designed to predict heart disease. It leverages the Flower (FLwr) framework to facilitate decentralized model training. Key components include a Keras Feedforward Neural Network (FFNN) model, a custom CifarClient class for interfacing with Flower, and functions for model training and evaluation. The client loads a subset of the heart disease dataset, updates its local FFNN model during training rounds, and contributes to the collaborative learning process. This client script is an integral part of a larger FL system and is intended to run on decentralized devices. The overarching goal is to collectively improve the FFNN model's accuracy while preserving data privacy.